Generalized loading assemblies
